requirements.txt
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0